fix: cleanup path before accessing shared data; new feature: return a directory listing when GETting a directory via share service

+++ authserver/src/main/java/eu/unicore/uftp/authserver/share/AccessServiceImpl.java
@@ -66,11 +66,16 @@ public class AccessServiceImpl extends ShareServiceBase {
 				|| equal(share.getTargetID(), Client.ANONYMOUS_CLIENT_DN)){
 				ShareServiceProperties spp = kernel.getAttribute(ShareServiceProperties.class);
 				String clientIP = spp.getClientIP();
-				handlePath(path, share);
-				logUsage(false, path, share);
-				r = doDownload(share, uftp, clientIP, range);
-				if(share.isOneTime()) {
-					shareDB.delete(uniqueID);
+				boolean returnListing = handlePath(path, share);
+				if(returnListing) {
+					r = getListing(share, uftp, clientIP);
+				}
+				else {
+					logUsage(false, path, share);
+					r = doDownload(share, uftp, clientIP, range);
+					if(share.isOneTime()) {
+						shareDB.delete(uniqueID);
+					}
 				}
 			}else{
 				throw new WebApplicationException(Status.UNAUTHORIZED);
@@ -81,21 +86,19 @@ public class AccessServiceImpl extends ShareServiceBase {
 		return r;
 	}
 
-	private void handlePath(String path, ShareDAO share) throws WebApplicationException{
+	private boolean handlePath(String path, ShareDAO share) throws WebApplicationException{
 		if(path!=null){
 			if(share.isDirectory()){
 				String fullPath = FilenameUtils.normalize(share.getPath()+"/"+path, true);
 				share.setPath(fullPath);
+				return false;
 			}
 			else{
 				throw new WebApplicationException("Not a directory", 404);
 			}
 		}
 		else{
-			if(share.isDirectory()){
-				// TODO we could return a listing
-				throw new WebApplicationException(404);
-			}
+			return share.isDirectory();
 		}
 	}
